+++ demo/src/app/sudoku/render/render_cell_value.rs
@@ -1,4 +1,4 @@
-use vertigo::{computed::Computed, VDomElement, Css, html, css, css_fn};
+use vertigo::{Computed, Css, VDomElement, html, css, css_fn};
 
 use crate::app::sudoku::state::{Cell, number_item::SudokuValue};